Simplify the regex abstraction by replacing the Java and RE2/J implementation classes with anonymous implementations.
