--- processing/src/main/java/org/apache/druid/regex/Re2jRegexMatcher.java
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.regex;
-
-import com.google.re2j.Matcher;
-
-public class Re2jRegexMatcher implements RegexMatcher
-{
-  private final Matcher matcher;
-
-  public Re2jRegexMatcher(Matcher matcher)
-  {
-    this.matcher = matcher;
-  }
-
-  @Override
-  public boolean matches()
-  {
-    return matcher.matches();
-  }
-
-  @Override
-  public boolean find()
-  {
-    return matcher.find();
-  }
-
-  @Override
-  public String group(int group)
-  {
-    return matcher.group(group);
-  }
-
-  @Override
-  public int groupCount()
-  {
-    return matcher.groupCount();
-  }
-}
