--- processing/src/main/java/org/apache/druid/regex/Re2jRegexPattern.java
@@ -1,38 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.regex;
-
-import com.google.re2j.Pattern;
-
-public class Re2jRegexPattern implements RegexPattern
-{
-  private final Pattern pattern;
-
-  public Re2jRegexPattern(String regex)
-  {
-    this.pattern = Pattern.compile(regex);
-  }
-
-  @Override
-  public RegexMatcher matcher(String input)
-  {
-    return new Re2jRegexMatcher(pattern.matcher(input));
-  }
-}
